+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Notification Leptos component.
 - Columns Leptos components.
 - Menu Leptos components.
+- A method to setup the library and generate the leptos-bulma.css file.
 
 ### Changed